% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmrd.R
\name{adj_overlap}
\alias{adj_overlap}
\title{Construct a GWAS overlap proportion matrix}
\usage{
adj_overlap(
  exposure_overlap_proportions,
  prop_gwas_overlap_Xs_and_Y,
  number_of_exposures
)
}
\arguments{
\item{exposure_overlap_proportions}{Scalar or matrix of overlap proportions between exposure GWAS.}

\item{prop_gwas_overlap_Xs_and_Y}{Scalar or vector of overlap proportions between exposures and outcome GWAS.}

\item{number_of_exposures}{Number of exposures.}
}
\value{
A named (p+1) x (p+1) matrix where rows/columns are labelled
\code{"Outcome"}, \code{"Exposure1"}, etc.
}
\description{
Returns the (p+1) x (p+1) matrix of pairwise GWAS sample overlap proportions
for use as the \code{prop_gwas_overlap_Xs} argument of \code{\link{set_params}}
when exposure GWAS samples partially overlap each other and the outcome GWAS.
}
\examples{
adj_overlap(
  exposure_overlap_proportions = 0.2,
  prop_gwas_overlap_Xs_and_Y = 0.1,
  number_of_exposures = 3
)
}
