% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{classIVs}
\alias{classIVs}
\title{Helper function}
\usage{
classIVs(ix, uhpix, chpix)
}
\arguments{
\item{ix}{ix}

\item{uhpix}{uhpix}

\item{chpix}{chpix}
}
\description{
Helper function
}
\examples{
\dontrun{
classIVs()
}
}
