% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{list_presets}
\alias{list_presets}
\title{List available simulation presets}
\usage{
list_presets()
}
\value{
Invisibly returns a named list of valid values.
}
\description{
Prints the valid values for each argument of \code{\link{load_preset}}.
}
\examples{
list_presets()
}
