% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{makeBlocks}
\alias{makeBlocks}
\title{A function to make LD blocks}
\usage{
makeBlocks(LD_causal_SNPs, number_of_causal_SNPs, nblocks = 1)
}
\arguments{
\item{LD_causal_SNPs}{the LD structure of the causal SNPs}

\item{number_of_causal_SNPs}{the total number of causal SNPs}

\item{nblocks}{the number of independent LD blocks}
}
\description{
Helper function to make LD blocks
}
\examples{
\dontrun{
makeBlocks()
}
}
