% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{parthcorr}
\alias{parthcorr}
\title{Helper function}
\usage{
parthcorr(x, n)
}
\arguments{
\item{x}{Matrix or numeric value}

\item{n}{Number of rows and columns of the matrix}
}
\description{
Helper function
}
\examples{
# parthcorr()
}
