% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmrd.R
\name{plot_simdata}
\alias{plot_simdata}
\title{Plot simulated data}
\usage{
plot_simdata(
  data,
  params = params,
  exposure_specific_plot = "total",
  verbose = TRUE
)
}
\arguments{
\item{data}{direct output from \code{generate()}}

\item{params}{Named list of parameters}

\item{exposure_specific_plot}{One of \code{'total'}, \code{'joint'}, \code{'conditional'}}

\item{verbose}{Logical, default \code{TRUE}}
}
\description{
Plot simulated data.
}
\examples{
\dontrun{
# If you used generate_summary(), execute the following
plot_simdata(gwas_data,summary_params) 

# If you used generate_individual(), execute the following
plot_simdata(gwas_data,individual_params) 
}
}
