% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{pruning}
\alias{pruning}
\title{Pruning SNPs}
\usage{
pruning(jointPs, R, r2)
}
\arguments{
\item{jointPs}{joint p-degree of freedom chi-square tests for IVs}

\item{R}{LD correlation matrix for SNPs}

\item{r2}{upper squared LD r2 threshold for pruning}
}
\description{
Helper function
}
\examples{
# pruning()
}
