% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{setf}
\alias{setf}
\title{Helper function}
\usage{
setf(bxunstd, nX, fix_Fstatistic_at)
}
\arguments{
\item{bxunstd}{bxunstd}

\item{nX}{nX}

\item{fix_Fstatistic_at}{Value to fix the F-statistic at}
}
\description{
Helper function
}
\examples{
\dontrun{
setf()
}
}
