% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{generate_individual}
\alias{generate_individual}
\title{An Individual-Level Data-Generating Function}
\usage{
generate_individual(params)
}
\arguments{
\item{params}{List of parameters used to generate simulated data}
}
\description{
This function generates simulated individual-level and data given a list of parameters
}
\examples{
\dontrun{
individual_params <-
  list(
    sample_size_Xs = 5e4, # exposure GWAS sample sizes
    sample_size_Y = 5e4, # outcome GWAS sample size
    prop_gwas_overlap_Xs_and_Y = 0.5, # proportion of exposure and outcome GWAS overlap
    number_of_exposures = 2, # 4 number of exposures
    phenotypic_correlation_Xs = 0.2, # phenotypic correlation between exposures
    genetic_correlation_Xs = 0, # genetic correlation between exposures
    Xs_variance_explained_by_U = 1/4 - 0.12, # exposures variance explained by confounder
    Y_variance_explained_by_Xs = c(0, 0.5), # outcome variance explained by exposures
    signs_of_causal_effects = c(1, 1), # signs of causal effects
    Y_variance_explained_by_U = 0.1, # outcome variance explained by confounder
    number_of_causal_SNPs = 200, # number of SNPs causing exposures
    mafs_of_causal_SNPs = stats::runif(100, 0.1, 0.5), # minor allele frequency of causal SNPs
    Xs_variance_explained_by_g = 0.12, # exposures variance explained by SNPs 
    number_of_UHP_causal_SNPs = 30, # number of UHP exposure SNPs
    number_of_CHP_causal_SNPs = 10, # number of CHP exposure SNPs
    Y_variance_explained_by_UHP = 0.05, # outcome variance explained by UHP SNPs
    U_variance_explained_by_CHP = 0.05, # outcome variance explained by CHP SNPs
    LD_causal_SNPs = 'I', # independent causal exposure SNPs
    number_of_LD_blocks = 1, # number of independent LD blocks
    MR_standardization = 'Z', # standardization of GWAS summary statistics 
    simtype = 'weak', # simulation performed using weak instruments
    MVMR_IV_selection_type = 'joint', # P-values for IV selection based on joint test for exposures
    IV_Pvalue_threshold = 1, # P-value threshold for candidate IVs
    LD_pruning_r2 = 1, # upper boundary of squared LD correlation
    N_of_LD_ref = Inf, # size of the LD reference panel
    fix_Fstatistic_at = 10 # average across exposures using full MVMR IV set
  )
gwas_data <- generate_individual(individual_params)
}
}
\keyword{data}
