% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{generate_individual}
\alias{generate_individual}
\title{An Individual-Level Data-Generating Function}
\usage{
generate_individual(params)
}
\arguments{
\item{params}{List of parameters used to generate simulated data}
}
\description{
This function generates simulated individual-level and data given a list of parameters
}
\examples{
# generate_individual()
}
\keyword{data}
