% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{generate_summary}
\alias{generate_summary}
\title{An Summary Statistic Data-Generating Function}
\usage{
generate_summary(params)
}
\arguments{
\item{params}{List of parameters used to generate simulated data}
}
\description{
This function generates simulated summary-level and data given a list of parameters
}
\examples{
\dontrun{
summary_params <-
  list(
    sample_size_Xs = 30000, # exposure GWAS sample sizes
    sample_size_Y = 30000, # outcome GWAS sample size
    prop_gwas_overlap_Xs_and_Y = 1, # proportion of exposures' and outcome GWAS overlap
    number_of_exposures = 3, # number of exposures
    number_of_causal_SNPs = 100, # number of SNPs causing each exposure
    number_of_UHP_causal_SNPs = 0, # number of UHP causal SNPs
    number_of_CHP_causal_SNPs = 20, # number of CHP causal SNPs
    ratio_of_UHP_variance = 0.15, # ratio of UHP variance to valid IV variance
    ratio_of_CHP_variance = 0.25, # ratio of CHP variance to valid IV variance
    CHP_correlation = -0.5, # correlation between CHP and valid IV effect sizes
    simtype = 'winners', # performs IV selection based on P-value
    fix_Fstatistic_at = 10, # ignored because simtype='winners'
    prop_gwas_overlap_Xs = 1, # overlap of exposures' GWAS
    phenotypic_correlation_Xs = 0.3, # phenotypic correlations between exposures
    genetic_correlation_Xs = 0.15,  # genetic correlation between exposures
    phenotypic_correlations_Xs_and_Y = 0.3, # phenotypic correlations b/w exposures and outcome
    true_causal_effects = 0.3, # true causal effect sizes
    Xs_variance_explained_by_g = 0.10, # exposure variance explained by SNPs
    LD_causal_SNPs = 'ar1(0.5)', # LD between causal exposure SNPs
    number_of_LD_blocks = 3, # number of independent LD blocks
    MR_standardization = 'none', # does not standardize GWAS estimates
    MVMR_IV_selection_type = 'union', # SNPs associated with >0 exposures are candidate IVs
    IV_Pvalue_threshold = 5e-8, # only SNPs with P<this threshold are candidate IVs
    LD_pruning_r2 = 1, # the upper LD r2 pruning threshold for candidate IVs
    N_of_LD_ref = Inf # the sample size of the LD reference panel
  )
gwas_data <- generate_summary(summary_params)
}
}
\keyword{data}
