% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{generate_summary}
\alias{generate_summary}
\title{An Summary Statistic Data-Generating Function}
\usage{
generate_summary(params)
}
\arguments{
\item{params}{List of parameters used to generate simulated data}
}
\description{
This function generates simulated summary-level and data given a list of parameters
}
\examples{
generate_summary()
}
\keyword{data}
