% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{plot_simdata}
\alias{plot_simdata}
\title{Helper function}
\usage{
plot_simdata(
  data,
  params = params,
  exposure_specific_plot = "total",
  verbose = TRUE
)
}
\arguments{
\item{data}{direct output from generate()}
}
\description{
Helper function
}
\examples{
# plot_simdata()
}
