% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr.R
\name{plot_simdata_lower}
\alias{plot_simdata_lower}
\title{Helper function}
\usage{
plot_simdata_lower(data, params = params, showFstat = TRUE)
}
\arguments{
\item{data}{direct output from \code{generate()}}

\item{params}{Named list of parameters}

\item{showFstat}{Logical, default \code{TRUE}}
}
\description{
Helper function
}
\examples{
\dontrun{
plot_simdata_lower()
}
}
